project/requirements.txt
Flask==3.0.3
pdfminer.six==20231228
python-docx==1.1.2
PyPDF2==3.0.1
